model_definitions/09398d14-c34f-4465-bcd1-cb4e34e20cee/model_modules/requirements.txt
numpy==1.16.1
pandas==1.0.5
scikit-learn==0.23.1
sklearn2pmml==0.60.0
jpmml-evaluator==0.3.1
